- Add protection against application hard-killing and attempt to restore database after one have taken place. - Add to PUT command usecase when blob key is sent after blob is successfully written. - If client waits response after blob is written always inform it in case if error occured during writing to database. - Force release of NetCache for 32-bit platform to use BerkeleyDB 4.6. - Eliminate compiler warnings. JIRA: CXX-760, CXX-765, CXX-779
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@39985 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/util/time_line.hpp
@@ -277,8 +277,8 @@ unsigned CTimeLine<BV>::x_TimeLineSlot(time_t tm) const
     if (tm <= m_TimeLineHead)
         return 0;
 
-    unsigned interval_head = (tm / m_DiscrFactor) * m_DiscrFactor;
-    unsigned diff = interval_head - m_TimeLineHead;
+    unsigned interval_head = (unsigned)((tm / m_DiscrFactor) * m_DiscrFactor);
+    unsigned diff = (unsigned)(interval_head - m_TimeLineHead);
     return diff / m_DiscrFactor;
 }
